fix: ping capabilities so it can work (#130)
fixes #99 

The error message for nix installed ping is:
```
❯ ping thinkpad
ping: socktype: SOCK_RAW
ping: socket: Operation not permitted
ping: => missing cap_net_raw+p capability or setuid?
```

Using `setuid` is out because that grants all root privileges. Went with
the missing capabilities.

Since Nix `ping` is a symlink, the capabilities have to be applied to
what the symlink is pointing to. A combination of `which` to find
`ping`, then `readlink` to find where `ping` symlinks points to, allows
for the use of `setcap` to grant the capabilities to the real `ping`
command.