fix: remove docker pass, secretservice credential providers (#131)
fixes #92 

Docker credential providers are named with prefix `docker-credential`.
Tab complete the names for the `pass` and `secretservice` providers.

Loop over the names, then delete the full path as returned by `which`.

This removal is not clean because the package should be removed, but
that should be fixed in the base image. When fixing an existing system,
a fixup is acceptable.